CI fix and move test
