+++ tests/worker/workflow_sandbox/test_restrictions.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import sys
 from dataclasses import dataclass
 from typing import ClassVar, Dict, Optional
 
@@ -10,7 +11,27 @@ from temporalio.worker.workflow_sandbox.restrictions import (
     RestrictionContext,
     SandboxMatcher,
     _RestrictedProxy,
+    _stdlib_module_names,
 )
+
+
+def test_workflow_sandbox_stdlib_module_names():
+    if sys.version_info != (3, 10):
+        pytest.skip("Test only runs on 3.10")
+    actual_names = ",".join(sorted(sys.stdlib_module_names))
+    # Uncomment to print code for generating these
+    code_lines = [""]
+    for mod_name in sorted(sys.stdlib_module_names):
+        if code_lines[-1]:
+            code_lines[-1] += ","
+        if len(code_lines[-1]) > 80:
+            code_lines.append("")
+        code_lines[-1] += mod_name
+    code = f'_stdlib_module_names = (\n    "' + '"\n    "'.join(code_lines) + '"\n)'
+    # TODO(cretz): Point releases may add modules :-(
+    assert (
+        actual_names == _stdlib_module_names
+    ), f"Expecting names as {actual_names}. In code as:\n{code}"
 
 
 @dataclass
